Week6 (#10)
* [WEEK6]: Objcet.equals 실습

* [WEEK6]: Objcet.hashcode 실습

* [WEEK6]: Bank Project Init

* [WEEK6] csv 파일 추가

* [WEEK6]: 전체 입출금 내역 조회

* [WEEK6]: 월별  입출금 내역 조회

* [WEEK6]: 데이터 파서 클래스 분리 -(SRP)적용

* [WEEK6]: 전체, 월별 입출금 내역 조회 메서드로 분리

* [WEEK6]: 결합도를 낮추기위해 parser Interface 적용

* [WEEK6]: Service 클래스로 계싼 로직 분리

* [WEEK6]: Analyzer 클래스 추가 후, main 로직 은닉

* [WEEK6] 커피머신 실습 코드 추가

* [WEEK6] 현재 주문 금액 추가, 출력 수정

* [WEEK6] 입력 포맷 안내 추가

* [WEEK6] 입력 검증 로직 추가

* [WEEK6] Logging,Log Level 설정

* [WEEK6] chore

* [WEEK6] log File로 출력

* [WEEK6]: logger xml 코드 추가

* feat: Logback 추가 (Slf4j을 이용해 구현체 변경 용이하게 변경)

* feat: Slf4j

* [WEEK6]: default method 추가

* [WEEK6]: Object로 사과와 바나나 담는 예제 추가

* [WEEK6]: K,V로 과일 담을수 있는 예제 Generic 활용

* [WEEK6]: BankState 경로 ㅜㅅ정

* add Functional Intergace

* [WEEK6] Junit 코드 추가

+++ week6/CoffeeShop/src/main/java/dev/bank/model/Order.java
@@ -9,9 +9,10 @@ import dev.bank.model.enums.OrderType;
 public class Order {
 	private List<Drink> drinkList = new ArrayList<>();
 	private OrderType orderType;
-	private int receivedAmount;
+	private long receivedAmount;
 
-	public Order(List<Drink> drinkList, OrderType orderType, int receivedAmount) {
+	public Order(List<Drink> drinkList, OrderType orderType, long receivedAmount) {
+
 		this.drinkList = drinkList;
 		this.orderType = orderType;
 		this.receivedAmount = receivedAmount;
@@ -25,7 +26,8 @@ public class Order {
 		return orderType;
 	}
 
-	public int getReceivedAmount() {
+	public long getReceivedAmount() {
+
 		return receivedAmount;
 	}
 }
